+++ packages/event-bus-local/src/initialize/index.ts
@@ -3,10 +3,10 @@ import { IEventBusService } from "@medusajs/types"
 
 export const initialize = async (): Promise<IEventBusService> => {
   const serviceKey = Modules.EVENT_BUS
-  const loaded = await MedusaModule.bootstrap(
+  const loaded = await MedusaModule.bootstrap<IEventBusService>(
     serviceKey,
     "@medusajs/event-bus-local"
   )
 
-  return loaded[serviceKey] as IEventBusService
+  return loaded[serviceKey]
 }