+++ packages/medusa/src/api/routes/admin/discounts/index.ts
@@ -209,7 +209,6 @@ export const defaultAdminDiscountsFields: (keyof Discount)[] = [
 ]
 
 export const defaultAdminDiscountsRelations = [
-  "rule",
   "parent_discount",
   "regions",
   "rule.conditions",