+++ packages/orchestration/src/__tests__/joiner/remote-joiner-data.ts
@@ -45,8 +45,9 @@ const fetchServiceDataCallback = async (
   relationship?: any
 ) => {
   const serviceConfig = expand.serviceConfig
-  const moduleRegistryName =
-    lowerCaseFirst(serviceConfig.serviceName) + "Service"
+  const moduleRegistryName = !serviceConfig.serviceName.endsWith("Service")
+    ? lowerCaseFirst(serviceConfig.serviceName) + "Service"
+    : serviceConfig.serviceName
 
   const service = container.resolve(moduleRegistryName)
   const methodName = relationship?.inverse
@@ -74,7 +75,7 @@ describe("RemoteJoiner", () => {
 
   it("Simple query of a service, its id and no fields specified", async () => {
     const query = {
-      service: "User",
+      service: "user",
       args: [
         {
           name: "id",
@@ -143,7 +144,7 @@ describe("RemoteJoiner", () => {
 
   it("Query of a service, expanding a property and restricting the fields expanded", async () => {
     const query = {
-      service: "User",
+      service: "user",
       args: [
         {
           name: "id",
@@ -215,7 +216,7 @@ describe("RemoteJoiner", () => {
 
   it("Query a service expanding multiple nested properties", async () => {
     const query = {
-      service: "Order",
+      service: "order",
       fields: ["number", "date", "products"],
       expands: [
         {