+++ packages/product/src/types/index.ts
@@ -3,3 +3,5 @@ import { IEventBusService } from "@medusajs/types"
 export type InitializeModuleInjectableDependencies = {
   eventBusService?: IEventBusService
 }
+
+export * from "./services"