+++ packages/types/src/product/service.ts
@@ -1,5 +1,6 @@
 import {
   CreateProductDTO,
+  UpdateProductDTO,
   FilterableProductCategoryProps,
   FilterableProductCollectionProps,
   FilterableProductProps,
@@ -11,11 +12,19 @@ import {
   ProductTagDTO,
   ProductVariantDTO,
 } from "./common"
-import { FindConfig } from "../common"
+
 import { Context } from "../shared-context"
+import { FindConfig } from "../common"
+import { JoinerServiceConfig } from "../joiner"
 
 export interface IProductModuleService {
-  retrieve(productId: string, sharedContext?: Context): Promise<ProductDTO>
+  __joinerConfig(): JoinerServiceConfig
+
+  retrieve(
+    productId: string,
+    config?: FindConfig<ProductDTO>,
+    sharedContext?: Context
+  ): Promise<ProductDTO>
 
   list(
     filters?: FilterableProductProps,
@@ -91,6 +100,11 @@ export interface IProductModuleService {
 
   create(
     data: CreateProductDTO[],
+    sharedContext?: Context
+  ): Promise<ProductDTO[]>
+
+  update(
+    data: UpdateProductDTO[],
     sharedContext?: Context
   ): Promise<ProductDTO[]>