+++ packages/oas/oas-github-ci/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @medusajs/oas-github-ci
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@0.2.11
+
 ## 1.0.10
 
 ### Patch Changes